chore(canon): keep type-only import fix within source-length limits
